Выносит генерацию overview в отдельный компонент
